Fix a warning.
git-svn-id: https://svn.filezilla-project.org/svn/FileZilla3/trunk@9173 ce1b9ea0-41a8-11dc-9b12-cb80f3b22977

+++ src/engine/directorylistingparser.cpp
@@ -228,7 +228,7 @@ public:
 		default:
 		case decimal:
 			if (m_number == std::numeric_limits<int64_t>::min()) {
-				constexpr int max = (std::numeric_limits<int64_t>::max() - 9) / 10;
+				constexpr int64_t max = (std::numeric_limits<int64_t>::max() - 9) / 10;
 				if (IsNumeric() || IsLeftNumeric()) {
 					m_number = 0;
 					for (unsigned int i = 0; i < m_len; ++i) {
@@ -262,7 +262,7 @@ public:
 			return m_number;
 		case hex:
 			{
-				constexpr int max = (std::numeric_limits<int64_t>::max() - 15) / 16;
+				constexpr int64_t max = (std::numeric_limits<int64_t>::max() - 15) / 16;
 				int64_t number = 0;
 				for (unsigned int i = 0; i < m_len; ++i) {
 					if (number > max) {
